refactor: split test class #685
